+++ pkg/docker/build.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/vertigobr/safira/pkg/execute"
 	"github.com/vertigobr/safira/pkg/utils"
+	"gopkg.in/gookit/color.v1"
 )
 
 type dockerBuild struct {
@@ -34,9 +35,9 @@ func Build(image, functionName, handler, language string, noCache bool, args map
 		taskBuildArgs := getTaskBuildArgs(db)
 
 		taskBuild := execute.Task{
-			Command:      "docker",
-			Args:         taskBuildArgs,
-			StreamStdio:  true,
+			Command:     "docker",
+			Args:        taskBuildArgs,
+			StreamStdio: true,
 		}
 
 		res, err := taskBuild.Execute()
@@ -50,7 +51,10 @@ func Build(image, functionName, handler, language string, noCache bool, args map
 
 		return nil
 	} else {
-		return fmt.Errorf("atributo lang da função %s é inválido", functionName)
+		return fmt.Errorf(
+			"%s The input from %s to lang is invalid, look at templates running: %s",
+			color.Red.Text("[!]"), language, color.Bold.Text("safira template list"),
+		)
 	}
 }
 
@@ -67,19 +71,19 @@ func createBuildFolder(functionName, handler, language string) (string, error) {
 
 	err := os.RemoveAll(buildPath)
 	if err != nil {
-		return "", fmt.Errorf("error ao limpar a pasta de build: %s\n", buildPath)
+		return "", fmt.Errorf("%s Error cleaning build folder: %s", color.Red.Text("[!]"), buildPath)
 	}
 
 	functionPath := path.Join(buildPath, "function")
 
 	err = utils.Copy(path.Join("./template/", language), buildPath, false, true)
 	if err != nil {
-		return "", fmt.Errorf("error ao copiar arquivos para build: %s - %s.\n", buildPath, err.Error())
+		return "", fmt.Errorf("%s Error copying template %s files to the build folder: %s", color.Red.Text("[!]"), language, buildPath)
 	}
 
 	err = utils.Copy(filepath.Clean(handler), filepath.Clean(functionPath), true, true)
 	if err != nil {
-		return "", fmt.Errorf("error ao copiar arquivos para build: %s - %s.\n", buildPath, err.Error())
+		return "", fmt.Errorf("%s Error copying function %s files to the build folder: %s", color.Red.Text("[!]"), functionName, buildPath)
 	}
 
 	return buildPath, nil