+++ pkg/get/download.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/vertigobr/safira/pkg/config"
 	"github.com/vertigobr/safira/pkg/git"
+	"gopkg.in/gookit/color.v1"
 )
 
 func downloadBinary(url, name string, binary bool) error {
@@ -21,7 +22,7 @@ func downloadBinary(url, name string, binary bool) error {
 
 	res, err := http.DefaultClient.Get(parsedURL.String())
 	if err != nil {
-		return fmt.Errorf("error ao obter conteúdo da requisição: %s", err.Error())
+		return fmt.Errorf("%s Error when obtaining request body: %s", color.Red.Text("[!]"), url)
 	}
 	defer res.Body.Close()
 
@@ -34,17 +35,17 @@ func downloadBinary(url, name string, binary bool) error {
 		// Criar arquivo
 		out, err := os.Create(fmt.Sprintf("%s/%s", dest, name))
 		if err != nil {
-			return fmt.Errorf("error ao criar arquivo %s: %s", name, err.Error())
+			return fmt.Errorf("%s Error when creating %s file", color.Red.Text("[!]"), name)
 		}
 		defer out.Close()
 
 		// Escreve o corpo da resposta no arquivo
 		if _, err := io.Copy(out, res.Body); err != nil {
-			return fmt.Errorf("error ao escrever conteúdo da requisição no arquivo %s: %s", name, err.Error())
+			return fmt.Errorf("%s Error writing request body to file %s", color.Red.Text("[!]"), name)
 		}
 
 		if err := os.Chmod(fmt.Sprintf("%s/%s", dest, name), 0700); err != nil {
-			return fmt.Errorf("error ao tornar o arquivo executavél - %s: %s", name, err.Error())
+			return fmt.Errorf("%s Error when making the X file an executable - %s", color.Red.Text("[!]"), name)
 		}
 	} else {
 		r := ioutil.NopCloser(res.Body)
@@ -65,7 +66,7 @@ func downloadBinary(url, name string, binary bool) error {
 	Gid, err := strconv.Atoi(u.Gid)
 
 	if err := os.Chown(fmt.Sprintf("%s/%s", dest, name), Uid, Gid); err != nil {
-		return fmt.Errorf("error ao mudar o dono da pasta de root para o usuário: %s", err.Error())
+		return fmt.Errorf("%s Error changing the owner of the root folder for the user", color.Red.Text("[!]"))
 	}
 
 	return nil